refactor(components, utils): move code for clarity
